apply jpa

+++ src/main/java/co/istad/chhaya/fswd_sbapp/dto/CreateProductRequest.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 
 public record CreateProductRequest(
         String name,
-        BigDecimal price
+        double price
 ){
 }
